requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
gunicorn==20.1.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
cloudinary==1.28.1
django-cloudinary-storage==0.3.0
Pillow==9.0.0